pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "anb-fastapi-app"
version = "0.1.0"
description = "FastAPI application for ANB Rising Stars project"
authors = [{ name = "ANB Team" }]
requires-python = ">=3.11"

dependencies = [
    "fastapi==0.119.0",
    "uvicorn==0.37.0",
    "SQLAlchemy==2.0.44",
    "psycopg2-binary==2.9.11",
    "pydantic==2.12.0",
    "python-dotenv==1.1.1",
    "redis==6.4.0",
    "celery==5.5.3",
    "passlib==1.7.4",
    "python-jose==3.5.0",
    "python-multipart==0.0.20",
    "PyYAML==6.0.3",
    "httpx==0.27.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools]
include-package-data = true
